+++ src/components/common/accordion/accordion.tsx
@@ -52,7 +52,7 @@ export const Accordion: React.FC = () => {
                   className="flex justify-between gap-8 w-full"
                   onClick={() => togglePanels({ ...panel, key: index })}
                 >
-                  <p className="text-left">{faq.question}</p>
+                  <p className="text-left xl:w-[750px]">{faq.question}</p>
                   <Arrow
                     className={clsx(
                       'min-w-4 min-h-4 transform transition-transform duration-300',
