+++ src/hooks/use-device-type.tsx
@@ -1,0 +1,30 @@
+import { useState, useEffect } from 'react';
+
+export const useDeviceType = () => {
+  const [deviceType, setDeviceType] = useState({
+    isDesktop: false,
+    isTablet: false,
+    isMobile: false,
+  });
+
+  useEffect(() => {
+    const handleResize = () => {
+      const width = window.innerWidth;
+
+      setDeviceType({
+        isDesktop: width >= 1280,
+        isTablet: width >= 768 && width < 1280,
+        isMobile: width < 768,
+      });
+    };
+
+    handleResize();
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return deviceType;
+};
