--- src/hooks/use-is-desctop.tsx
@@ -1,20 +1,0 @@
-import { useState, useEffect } from 'react';
-
-export const useIsDesktop = () => {
-  const [isDesktop, setIsDesktop] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsDesktop(window.innerWidth >= 1280);
-    };
-
-    handleResize();
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
-
-  return isDesktop;
-};
