--- src/hooks/use-is-tablet.tsx
@@ -1,20 +1,0 @@
-import { useState, useEffect } from 'react';
-
-export const useIsTablet = () => {
-  const [isTablet, setIsTablet] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsTablet(window.innerWidth >= 768 && window.innerWidth < 1280);
-    };
-
-    handleResize();
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
-
-  return isTablet;
-};
