+++ src/sections/faq/faq.tsx
@@ -5,7 +5,7 @@ import { Accordion } from '@/components/common/accordion/accordion';
 export const FAQ = () => {
   return (
     <section className="bg-pinkSupperLight py-10 md:py-[60px] xl:py-[70px] ">
-      <div className="container">
+      <div className="container xl:w-[858px]">
         <h2 className="mb-6 md:mb-11 xl:text-center">{faq.title}</h2>
         <Accordion />
       </div>
